# openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio Dashboard API
  version: '0.1.0'
  description: Minimal API for the portfolio dashboard (LAN deployment)
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /finance:
    get:
      tags: [Finance]
      summary: Export finance JSON state
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Finance JSON export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceJSON'
  /finance/import:
    post:
      tags: [Finance]
      summary: Import finance JSON state
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinanceJSON'
      responses:
        '201':
          description: Imported
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
...
  /finance/incomes:
    post:
      tags: [Finance - Income]
      summary: Add income
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncomeCreate'
      responses:
        '201':
          description: Income added
          content:
            application/json:
              schema:
                type: object
                properties:
                  income:
                    $ref: '#/components/schemas/Income'
  /finance/incomes/{id}:
    put:
      tags: [Finance - Income]
      summary: Update income
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncomeUpdate'
      responses:
        '200':
          description: Income updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  income:
                    $ref: '#/components/schemas/Income'
    delete:
      tags: [Finance - Income]
      summary: Delete income
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
  /finance/fixed:
    get:
      tags: [Finance - Fixed]
      summary: List fixed expenses
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Fixed expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FixedExpense'
    post:
      tags: [Finance - Fixed]
      summary: Add fixed expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FixedExpenseCreate'
      responses:
        '201':
          description: Fixed expense added
          content:
            application/json:
              schema:
                type: object
                properties:
                  fixed:
                    $ref: '#/components/schemas/FixedExpense'
  /finance/fixed/{id}:
    put:
      tags: [Finance - Fixed]
      summary: Update fixed expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FixedExpenseUpdate'
      responses:
        '200':
          description: Fixed expense updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  fixed:
                    $ref: '#/components/schemas/FixedExpense'
    delete:
      tags: [Finance - Fixed]
      summary: Delete fixed expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
  /finance/eventual:
    get:
      tags: [Finance - Eventual]
      summary: List eventual expenses
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Eventual expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventualExpense'
    post:
      tags: [Finance - Eventual]
      summary: Add eventual expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventualExpenseCreate'
      responses:
        '201':
          description: Eventual expense added
          content:
            application/json:
              schema:
                type: object
                properties:
                  eventual:
                    $ref: '#/components/schemas/EventualExpense'
  /finance/eventual/{id}:
    put:
      tags: [Finance - Eventual]
      summary: Update eventual expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventualExpenseUpdate'
      responses:
        '200':
          description: Eventual expense updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  eventual:
                    $ref: '#/components/schemas/EventualExpense'
    delete:
      tags: [Finance - Eventual]
      summary: Delete eventual expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
  /finance/credit:
    get:
      tags: [Finance - Credit]
      summary: List credit card expenses
      parameters:
        - name: card
          in: query
          required: true
          schema:
            type: string
            enum: [nuRenan, nuJu, nomad]
        - name: month
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Credit expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreditExpense'
    post:
      tags: [Finance - Credit]
      summary: Add credit expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditExpenseCreate'
      responses:
        '201':
          description: Credit expense added
          content:
            application/json:
              schema:
                type: object
                properties:
                  credit:
                    $ref: '#/components/schemas/CreditExpense'
  /finance/credit/{id}:
    put:
      tags: [Finance - Credit]
      summary: Update credit expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditExpenseUpdate'
      responses:
        '200':
          description: Credit expense updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  credit:
                    $ref: '#/components/schemas/CreditExpense'
    delete:
      tags: [Finance - Credit]
      summary: Delete credit expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
  /finance/years:
    get:
      tags: [Finance - Years]
      summary: List available finance years
      responses:
        '200':
          description: Years list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
    post:
      tags: [Finance - Years]
      summary: Create a finance year
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/YearPayload'
      responses:
        '201':
          description: Year created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  year:
                    type: integer
  /finance/years/{year}/reset:
    post:
      tags: [Finance - Years]
      summary: Reset finance data for a year
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Year reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  year:
                    type: integer
components:
  schemas:
...
    Income:
      type: object
      properties:
        id:
          type: string
        month:
          type: integer
        year:
          type: integer
        source:
          type: string
        value:
          type: number
        description:
          type: string
          nullable: true
        ts:
          type: integer
    IncomeCreate:
      type: object
      required: [month, source, value]
      properties:
        month:
          type: integer
        source:
          type: string
        value:
          type: number
        description:
          type: string
        year:
          type: integer
    IncomeUpdate:
      type: object
      required: [source, value]
      properties:
        source:
          type: string
        value:
          type: number
        description:
          type: string
    FixedExpense:
      type: object
      properties:
        id:
          type: string
        month:
          type: integer
        year:
          type: integer
        name:
          type: string
        category:
          type: string
        value:
          type: number
        paymentMethod:
          type: string
          nullable: true
        normallyDueDay:
          type: integer
          nullable: true
        paidOnDate:
          type: string
          nullable: true
        ts:
          type: integer
    FixedExpenseCreate:
      type: object
      required: [month, name, category, value]
      properties:
        month:
          type: integer
        name:
          type: string
        category:
          type: string
        value:
          type: number
        paymentMethod:
          type: string
        normallyDueDay:
          type: integer
        paidOnDate:
          type: string
        year:
          type: integer
    FixedExpenseUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        value:
          type: number
        paymentMethod:
          type: string
        normallyDueDay:
          type: integer
          nullable: true
        paidOnDate:
          type: string
          nullable: true
    EventualExpense:
      type: object
      properties:
        id:
          type: string
        month:
          type: integer
        year:
          type: integer
        category:
          type: string
        value:
          type: number
        description:
          type: string
          nullable: true
        paymentMethod:
          type: string
          nullable: true
        ts:
          type: integer
    EventualExpenseCreate:
      type: object
      required: [month, category, value]
      properties:
        month:
          type: integer
        category:
          type: string
        value:
          type: number
        description:
          type: string
        paymentMethod:
          type: string
        year:
          type: integer
    EventualExpenseUpdate:
      type: object
      properties:
        category:
          type: string
        value:
          type: number
        description:
          type: string
        paymentMethod:
          type: string
    CreditExpense:
      type: object
      properties:
        id:
          type: string
        card:
          type: string
        month:
          type: integer
        year:
          type: integer
        category:
          type: string
        value:
          type: number
        paymentMethod:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        paidOnDate:
          type: string
          nullable: true
        ts:
          type: integer
    CreditExpenseCreate:
      type: object
      required: [card, month, category, value]
      properties:
        card:
          type: string
          enum: [nuRenan, nuJu, nomad]
        month:
          type: integer
        category:
          type: string
        value:
          type: number
        paymentMethod:
          type: string
        description:
          type: string
        year:
          type: integer
    CreditExpenseUpdate:
      type: object
      properties:
        category:
          type: string
        value:
          type: number
        paymentMethod:
          type: string
        description:
          type: string
        paidOnDate:
          type: string
          nullable: true
    YearPayload:
      type: object
      required: [year]
      properties:
        year:
          type: integer
    FinanceJSON:
      type: object
      properties:
        year:
          type: integer
        incomes:
          type: array
          items:
            type: array
            items:
              type: object
        tithes:
          type: array
          items:
            type: array
            items:
              type: object
        fixedExpenses:
          type: array
          items:
            type: array
            items:
              type: object
        eventualExpenses:
          type: array
          items:
            type: array
            items:
              type: object
        creditCards:
          type: object
          properties:
            nuRenan:
              type: array
              items:
                type: array
                items:
                  type: object
            nuJu:
              type: array
              items:
                type: array
                items:
                  type: object
            nomad:
              type: array
              items:
                type: array
                items:
                  type: object